#28 Refactor 'adapters' package name
